pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = ""
authors = ["coder-sabarish <studysabarish@gmail.com>"]
readme = "README.md"
packages = [{include = "todo_app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.103.2"}
sqlalchemy = "^2.0.22"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pydocstyle = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
